backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
pdfplumber==0.10.3
scikit-learn==1.3.2
pydantic==2.5.2
